Timer Long 타입 수정

+++ BackEnd/src/main/java/com/watch/switme/controller/TimerController.java
@@ -69,7 +69,7 @@ public class TimerController {
         Study study=timer.getStudy();
 
         if(study!=null){
-            int study_idx=study.getStudy_idx();
+            Long study_idx=study.getStudy_idx();
         }
